Rename createScene.js to Game.js
