+++ src/index.js
@@ -1,7 +1,6 @@
 import * as BABYLON from "@babylonjs/core/Legacy/legacy";
 
-import CreateScene from "./CreateScene.js";
-import Game from "./CreateScene.js";
+import Game from "./Game.js";
 
 let canvas = document.getElementById("renderCanvas"); // Get the canvas element
 let engine = new BABYLON.Engine(canvas, true); // Generate the BABYLON 3D engine
